move jobs and listeners to its corresponding namespaces
