+++ app/integrations/discord/listeners/create_external_message_listener.rb
@@ -1,0 +1,9 @@
+module Discord
+  module Listeners
+    class CreateExternalMessageListener
+      def discord_engine_message_created(payload)
+        DiscordMessages::Creator.new(params: payload[:response], message_id: payload[:reference_id]).call
+      end
+    end
+  end
+end
