--- app/listeners/create_external_message_listener.rb
@@ -1,5 +1,0 @@
-class CreateExternalMessageListener
-  def discord_engine_message_created(payload)
-    DiscordMessages::Creator.new(params: payload[:response], message_id: payload[:reference_id]).call
-  end
-end
