+++ app/models/interface.rb
@@ -1,7 +1,6 @@
 class Interface < ApplicationRecord
-  DISCORD_GUILD_SOURCE = :discord_guild
   INTERNAL_SOURCES = %i[web api].freeze
-  EXTERNAL_SOURCES = [DISCORD_GUILD_SOURCE].freeze
+  EXTERNAL_SOURCES = [:discord_guild].freeze
 
   enum source: { discord_guild: 0, web: 1, api: 2 }
