+++ config/initializers/listeners.rb
@@ -1,7 +1,8 @@
 Rails.application.reloader.to_prepare do
   Wisper.clear if Rails.env.development?
 
-  Wisper.subscribe(CreateExternalMessageListener.new)
-  Wisper.subscribe(PrivyMessageExpiredListener.new)
-  Wisper.subscribe(MessageReadListener.new)
+  Wisper.subscribe(Discord::Listeners::CreateExternalMessageListener.new)
+  Wisper.subscribe(Discord::Listeners::SendExpiredNoticeListener.new)
+  Wisper.subscribe(Discord::Listeners::HideMessageListener.new)
+  Wisper.subscribe(Discord::Listeners::ExpireMessageListener.new)
 end
